feat: update status aduan warga

+++ app/Http/Controllers/ComplaintController.php
@@ -6,13 +6,16 @@ use App\Models\Complaint;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class ComplaintController extends Controller
 {
     public function index()
     {
         $residentId = Auth::user()->resident->id ?? null;
-        $complaints = Complaint::where('resident_id', $residentId)->paginate(5);
+        $complaints = Complaint::when(Auth::user()->role_id == 2, function ($query) use ($residentId) {
+            $query->where('resident_id', $residentId);
+        })->paginate(5);
 
         return view('pages.complaint.index', compact(
             'complaints',
@@ -83,6 +86,11 @@ class ComplaintController extends Controller
         }
 
         $complaint = Complaint::findOrFail($id);
+
+        if ($complaint->status !== 'new') {
+            return redirect('/complaint')->with('error', "Gagal menghapus aduan, status anda saat ini adalah $complaint->status_label");
+        }
+
         $complaint->resident_id = $resident->id;
         $complaint->title = $request->input('title');
         $complaint->content = $request->input('content');
@@ -100,6 +108,25 @@ class ComplaintController extends Controller
         return redirect('/complaint')->with('success', 'Behasil mengubah aduan');
     }
 
+    public function update_status(Request $request, $id)
+    {
+        $request->validate([
+            'status' => ['required', Rule::in(['new', 'processing', 'completed'])],
+        ]);
+
+        $resident = Auth::user()->resident;
+        if (Auth::user()->role_id == 2 && !$resident) {
+            return redirect('/complaint')->with('error', 'Akun anda belum terhubung dengan data penduduk');
+        }
+
+        $complaint = Complaint::findOrFail($id);
+        $complaint->status = $request->input('status');
+
+        $complaint->save();
+
+        return redirect('/complaint')->with('success', 'Behasil mengubah status');
+    }
+
     public function destroy($id)
     {
         $resident = Auth::user()->resident;
@@ -107,6 +134,9 @@ class ComplaintController extends Controller
             return redirect('/complaint')->with('error', 'Akun anda belum terhubung dengan data penduduk');
         }
         $complaint = Complaint::findOrFail($id);
+        if ($complaint->status !== 'new') {
+            return redirect('/complaint')->with('error', "Gagal menghapus aduan, status anda saat ini adalah $complaint->status_label");
+        }
         $complaint->delete();
 
         return redirect('/complaint')->with('success', 'Behasil menghapus aduan');
